# kotlin/rayfold-core/src/test/resources/oracle/bookstore.openapi.prefixed.json
{
 "openapi": "3.2.0",
 "info": {
  "title": "Bookstore",
  "version": "2.1"
 },
 "paths": {
...
  "/api/books/{id}": {
   "get": {
    "operationId": "book",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "shape",
      "in": "query",
      "required": false,
      "description": "Rayfold shape text or sha256: shape id; default view when absent",
      "schema": {
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Result in the requested shape (default view when no shape is given)",
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/Book"
          },
          {
           "type": "null"
          }
         ]
        }
       }
      }
     },
     "304": {
      "description": "Not modified (ETag revalidation)"
     },
     "400": {
      "description": "Invalid argument, including schema constraints such as @range",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "401": {
      "description": "Sign-in required by a policy",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "403": {
      "description": "Denied by a policy",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     }
    }
   },
   "patch": {
    "operationId": "updateBook",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "summary": "Change some fields of a book; fields not sent are left as they are.",
    "requestBody": {
     "required": true,
     "content": {
      "application/merge-patch+json": {
       "schema": {
        "$ref": "#/components/schemas/BookPatch"
       }
      },
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/BookPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Result in the requested shape (default view when no shape is given)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Book"
        }
       }
      }
     },
     "400": {
      "description": "Invalid argument, including schema constraints such as @range",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "401": {
      "description": "Sign-in required by a policy",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "403": {
      "description": "Denied by a policy",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Page_Book": {
    "type": "object",
    "properties": {
     "items": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Book"
      }
     },
     "cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     },
     "hasMore": {
      "type": "boolean"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "additionalProperties": false,
    "required": [
     "items",
     "hasMore"
    ]
   },
   "Book": {
    "type": "object",
    "properties": {
     "$type": {
      "const": "Book"
     },
     "id": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "format": {
      "type": "string",
      "enum": [
       "HARDCOVER",
       "PAPERBACK",
       "EBOOK"
      ]
     },
     "price": {
      "type": "string",
      "pattern": "^-?\\d+(\\.\\d+)?$",
      "description": "Unit price in the store currency."
     },
     "stock": {
      "type": "integer"
     },
     "author": {
      "$ref": "#/components/schemas/Author"
     },
     "reviews": {
      "$ref": "#/components/schemas/Page_Review"
     },
     "costPrice": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^-?\\d+(\\.\\d+)?$"
       },
       {
        "type": "null"
       }
      ],
      "description": "Only the owner (or an admin) can see the cost price."
     },
     "ownerId": {
      "type": "string"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "title",
     "format",
     "price",
     "stock",
     "author",
     "reviews",
     "ownerId"
    ]
   },
   "Author": {
    "type": "object",
    "properties": {
     "$type": {
      "const": "Author"
     },
     "id": {
      "type": "string"
     },
     "name": {
      "type": "string"
     },
     "bio": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     },
     "books": {
      "$ref": "#/components/schemas/Page_Book"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "name",
     "books"
    ]
   },
   "Page_Review": {
    "type": "object",
    "properties": {
     "items": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Review"
      }
     },
     "cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     },
     "hasMore": {
      "type": "boolean"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "additionalProperties": false,
    "required": [
     "items",
     "hasMore"
    ]
   },
   "Review": {
    "type": "object",
    "properties": {
     "$type": {
      "const": "Review"
     },
     "id": {
      "type": "string"
     },
     "rating": {
      "type": "integer",
      "x-rayfold-range": {
       "min": 1,
       "max": 5
      },
      "minimum": 1,
      "maximum": 5
     },
     "body": {
      "type": "string"
     },
     "book": {
      "$ref": "#/components/schemas/Book"
     },
     "reviewerId": {
      "type": "string"
     },
     "version": {
      "type": "integer",
      "description": "Bumped on every edit; send it back as ifVersion (or If-Match) to avoid lost updates."
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "rating",
     "body",
     "book",
     "reviewerId",
     "version"
    ]
   },
...
   "BookPatch": {
    "type": "object",
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "x-rayfold-range": {
       "min": 1,
       "max": 200
      },
      "minLength": 1,
      "maxLength": 200
     },
     "price": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^-?\\d+(\\.\\d+)?$"
       },
       {
        "type": "null"
       }
      ],
      "x-rayfold-range": {
       "min": 0
      }
     },
     "stock": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "x-rayfold-range": {
       "min": 0
      },
      "minimum": 0
     }
    },
    "additionalProperties": false,
    "description": "Partial update: absent fields are left unchanged."
   },
   "Problem": {
    "type": "object",
    "description": "RFC 9457 problem details; Rayfold errors keep code, type and data",
    "properties": {
     "type": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "status": {
      "type": "integer"
     },
     "detail": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "path": {
      "type": "string"
     },
     "data": {}
    },
    "required": [
     "type",
     "title",
     "status",
     "code"
    ]
   }
  }
 }
}